REFACTOR::  move run_with_timeout to utils script
